requirements.txt
numpy==1.26.4
pandas==2.1.4
scapy==2.5.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.30.2
sentence-transformers==2.2.2
huggingface-hub==0.15.1
tokenizers==0.13.3
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
networkx==3.2.1
tqdm==4.66.1
tensorflow-cpu==2.15.0
tf-keras==2.15.0